Sailthru Cleanup
- Moved tasks to end of file, below the functions consumed by the tasks
- Added helper functions to get Sailthru configuration and API client
- Added exception classes

ECOM-6975
