+++ ecommerce_worker/sailthru/v1/tasks.py
@@ -4,106 +4,20 @@ This file contains celery tasks for email marketing signal handler.
 
 from celery import shared_task
 from celery.utils.log import get_task_logger
-from sailthru.sailthru_client import SailthruClient
 from sailthru.sailthru_error import SailthruClientError
 
 from ecommerce_worker.cache import Cache
-from ecommerce_worker.utils import get_configuration, get_ecommerce_client
+from ecommerce_worker.sailthru.v1.exceptions import SailthruError
+from ecommerce_worker.sailthru.v1.utils import get_sailthru_client, get_sailthru_configuration
+from ecommerce_worker.utils import get_ecommerce_client
 
 logger = get_task_logger(__name__)  # pylint: disable=invalid-name
 cache = Cache()  # pylint: disable=invalid-name
 
 
-# pylint: disable=not-callable
-@shared_task(bind=True, ignore_result=True)
-def update_course_enrollment(self, email, course_url, purchase_incomplete, mode,
-                             unit_cost=None, course_id=None,
-                             currency=None, message_id=None, site_code=None):  # pylint: disable=unused-argument
-    """Adds/updates Sailthru when a user adds to cart/purchases/upgrades a course
+# pylint: disable=unused-argument
 
-     Args:
-        email(str): The user's email address
-        course_url(str): Course home page url
-        incomplete(boolean): True if adding to cart
-        mode(string): enroll mode (audit, verification, ...)
-        unit_cost(decimal): cost if purchase event
-        course_id(CourseKey): course id
-        currency(str): currency if purchase event - currently ignored since Sailthru only supports USD
-        message_id(str): value from Sailthru marketing campaign cookie
-        site_code(str): site code
-
-    Returns:
-        None
-    """
-    # Get configuration
-    config = get_configuration('SAILTHRU', site_code=site_code)
-
-    # Return if Sailthru integration disabled
-    if not config.get('SAILTHRU_ENABLE'):
-        return
-
-    # Make sure key and secret configured
-    sailthru_key = config.get('SAILTHRU_KEY')
-    sailthru_secret = config.get('SAILTHRU_SECRET')
-    if not (sailthru_key and sailthru_secret):
-        logger.error("Sailthru key and or secret not specified for site %s", site_code)
-        return
-
-    sailthru_client = SailthruClient(sailthru_key, sailthru_secret)
-
-    # Use event type to figure out processing required
-    new_enroll = False
-    send_template = None
-
-    if not purchase_incomplete:
-        if mode == 'verified':
-            # upgrade complete
-            send_template = config.get('SAILTHRU_UPGRADE_TEMPLATE')
-        elif mode == 'audit' or mode == 'honor':
-            # free enroll
-            new_enroll = True
-            send_template = config.get('SAILTHRU_ENROLL_TEMPLATE')
-        else:
-            # paid course purchase complete
-            new_enroll = True
-            send_template = config.get('SAILTHRU_PURCHASE_TEMPLATE')
-
-    # calc price in pennies for Sailthru
-    #  https://getstarted.sailthru.com/new-for-developers-overview/advanced-features/purchase/
-    cost_in_cents = int(unit_cost * 100)
-    if not cost_in_cents:
-        cost_in_cents = config.get('SAILTHRU_MINIMUM_COST')
-        # if still zero, ignore purchase since Sailthru can't deal with $0 transactions
-        if not cost_in_cents:
-            return
-
-    # update the "unenrolled" course array in the user record on Sailthru if new enroll or unenroll
-    if new_enroll:
-        if not _update_unenrolled_list(sailthru_client, email, course_url, False):
-            _schedule_retry(self, config)
-
-    # Get course data from Sailthru content library or cache
-    course_data = _get_course_content(course_id, course_url, sailthru_client, site_code, config)
-
-    # build item description
-    item = _build_purchase_item(course_id, course_url, cost_in_cents, mode, course_data)
-
-    # build purchase api options list
-    options = {}
-    if purchase_incomplete and config.get('SAILTHRU_ABANDONED_CART_TEMPLATE'):
-        options['reminder_template'] = config.get('SAILTHRU_ABANDONED_CART_TEMPLATE')
-        # Sailthru reminder time format is '+n time unit'
-        options['reminder_time'] = "+{} minutes".format(config.get('SAILTHRU_ABANDONED_CART_DELAY'))
-
-    # add appropriate send template
-    if send_template:
-        options['send_template'] = send_template
-
-    if not _record_purchase(sailthru_client, email, item, purchase_incomplete, message_id, options):
-        _schedule_retry(self, config)
-
-
-def _schedule_retry(self, config):
+def schedule_retry(self, config):
     """Schedule a retry"""
     raise self.retry(countdown=config.get('SAILTHRU_RETRY_SECONDS'),
                      max_retries=config.get('SAILTHRU_RETRY_ATTEMPTS'))
@@ -158,7 +72,7 @@ def _record_purchase(sailthru_client, email, item, purchase_incomplete, message_
         if not sailthru_response.is_ok():
             error = sailthru_response.get_error()
             logger.error("Error attempting to record purchase in Sailthru: %s", error.get_message())
-            return not _retryable_sailthru_error(error)
+            return not can_retry_sailthru_request(error)
 
     except SailthruClientError as exc:
         logger.exception("Exception attempting to record purchase for %s in Sailthru - %s", email, unicode(exc))
@@ -254,13 +168,13 @@ def _update_unenrolled_list(sailthru_client, email, course_url, unenroll):
         if not sailthru_response.is_ok():
             error = sailthru_response.get_error()
             logger.error("Error attempting to read user record from Sailthru: %s", error.get_message())
-            return not _retryable_sailthru_error(error)
+            return not can_retry_sailthru_request(error)
 
         response_json = sailthru_response.json
 
         unenroll_list = []
         if response_json and "vars" in response_json and response_json["vars"] \
-                and "unenrolled" in response_json["vars"]:
+           and "unenrolled" in response_json["vars"]:
             unenroll_list = response_json["vars"]["unenrolled"]
 
         changed = False
@@ -283,7 +197,7 @@ def _update_unenrolled_list(sailthru_client, email, course_url, unenroll):
             if not sailthru_response.is_ok():
                 error = sailthru_response.get_error()
                 logger.error("Error attempting to update user record in Sailthru: %s", error.get_message())
-                return not _retryable_sailthru_error(error)
+                return not can_retry_sailthru_request(error)
 
         return True
 
@@ -292,14 +206,101 @@ def _update_unenrolled_list(sailthru_client, email, course_url, unenroll):
         return False
 
 
-def _retryable_sailthru_error(error):
-    """ Return True if error should be retried.
+def can_retry_sailthru_request(error):
+    """ Returns True if a Sailthru request and be re-submitted after an error has occurred.
 
-    9: Retryable internal error
-    43: Rate limiting response
-    others: Not retryable
+    Responses with the following codes can be retried:
+         9: Internal Error
+        43: Too many [type] requests this minute to /[endpoint] API
 
-    See: https://getstarted.sailthru.com/new-for-developers-overview/api/api-response-errors/
+    All other errors are considered failures, that should not be retried. A complete list of error codes is available at
+    https://getstarted.sailthru.com/new-for-developers-overview/api/api-response-errors/.
+
+    Args:
+        error (SailthruResponseError)
+
+    Returns:
+        bool: Indicates if the original request can be retried.
     """
     code = error.get_error_code()
-    return code == 9 or code == 43
+    return code in (9, 43)
+
+
+@shared_task(bind=True, ignore_result=True)
+def update_course_enrollment(self, email, course_url, purchase_incomplete, mode, unit_cost=None, course_id=None,
+                             currency=None, message_id=None, site_code=None):
+    """Adds/updates Sailthru when a user adds to cart/purchases/upgrades a course
+
+     Args:
+        email(str): The user's email address
+        course_url(str): Course home page url
+        purchase_incomplete(boolean): True if adding to cart
+        mode(string): enroll mode (audit, verification, ...)
+        unit_cost(decimal): cost if purchase event
+        course_id(CourseKey): course id
+        currency(str): currency if purchase event - currently ignored since Sailthru only supports USD
+        message_id(str): value from Sailthru marketing campaign cookie
+        site_code(str): site code
+
+    Returns:
+        None
+    """
+    # Get configuration
+    config = get_sailthru_configuration(site_code)
+
+    try:
+        sailthru_client = get_sailthru_client(site_code)
+    except SailthruError:
+        # NOTE: We rely on the function to log the error for us
+        return
+
+    # Use event type to figure out processing required
+    new_enroll = False
+    send_template = None
+
+    if not purchase_incomplete:
+        if mode == 'verified':
+            # upgrade complete
+            send_template = config.get('SAILTHRU_UPGRADE_TEMPLATE')
+        elif mode == 'audit' or mode == 'honor':
+            # free enroll
+            new_enroll = True
+            send_template = config.get('SAILTHRU_ENROLL_TEMPLATE')
+        else:
+            # paid course purchase complete
+            new_enroll = True
+            send_template = config.get('SAILTHRU_PURCHASE_TEMPLATE')
+
+    # calc price in pennies for Sailthru
+    #  https://getstarted.sailthru.com/new-for-developers-overview/advanced-features/purchase/
+    cost_in_cents = int(unit_cost * 100)
+    if not cost_in_cents:
+        cost_in_cents = config.get('SAILTHRU_MINIMUM_COST')
+        # if still zero, ignore purchase since Sailthru can't deal with $0 transactions
+        if not cost_in_cents:
+            return
+
+    # update the "unenrolled" course array in the user record on Sailthru if new enroll or unenroll
+    if new_enroll:
+        if not _update_unenrolled_list(sailthru_client, email, course_url, False):
+            schedule_retry(self, config)
+
+    # Get course data from Sailthru content library or cache
+    course_data = _get_course_content(course_id, course_url, sailthru_client, site_code, config)
+
+    # build item description
+    item = _build_purchase_item(course_id, course_url, cost_in_cents, mode, course_data)
+
+    # build purchase api options list
+    options = {}
+    if purchase_incomplete and config.get('SAILTHRU_ABANDONED_CART_TEMPLATE'):
+        options['reminder_template'] = config.get('SAILTHRU_ABANDONED_CART_TEMPLATE')
+        # Sailthru reminder time format is '+n time unit'
+        options['reminder_time'] = "+{} minutes".format(config.get('SAILTHRU_ABANDONED_CART_DELAY'))
+
+    # add appropriate send template
+    if send_template:
+        options['send_template'] = send_template
+
+    if not _record_purchase(sailthru_client, email, item, purchase_incomplete, message_id, options):
+        schedule_retry(self, config)
